增加起始、终止地址

+++ oh-module/oh-urban/src/main/java/com/finn/urban/entity/TrajectoryEntity.java
@@ -18,18 +18,18 @@ public class TrajectoryEntity extends BaseEntity {
     /**
      * 位置的精确度，反应与真实位置之间的接近程度，数值越小越精确
      */
-    private BigDecimal accuracy;
+    private Double accuracy;
 
     /**
      * 高度，单位 m
      */
-    private BigDecimal altitude;
+    private Double altitude;
 
     /**
      * 水平精度，单位 m
      */
     @TableField("horizontal_accuracy")
-    private BigDecimal horizontalAccuracy;
+    private Double horizontalAccuracy;
 
     /**
      * 经度，范围为 -180~180，负数表示西经，使用 gcj02 国测局坐标系
@@ -44,13 +44,13 @@ public class TrajectoryEntity extends BaseEntity {
     /**
      * 速度，单位 m/s
      */
-    private BigDecimal speed;
+    private Double speed;
 
     /**
      * 垂直精度，单位 m
      */
     @TableField("vertical_accuracy")
-    private BigDecimal verticalAccuracy;
+    private Double verticalAccuracy;
 
     /**
      * gps定位时间
@@ -81,27 +81,27 @@ public class TrajectoryEntity extends BaseEntity {
      */
     private String remark;
 
-    public BigDecimal getAccuracy() {
+    public Double getAccuracy() {
         return accuracy;
     }
 
-    public void setAccuracy(BigDecimal accuracy) {
+    public void setAccuracy(Double accuracy) {
         this.accuracy = accuracy;
     }
 
-    public BigDecimal getAltitude() {
+    public Double getAltitude() {
         return altitude;
     }
 
-    public void setAltitude(BigDecimal altitude) {
+    public void setAltitude(Double altitude) {
         this.altitude = altitude;
     }
 
-    public BigDecimal getHorizontalAccuracy() {
+    public Double getHorizontalAccuracy() {
         return horizontalAccuracy;
     }
 
-    public void setHorizontalAccuracy(BigDecimal horizontalAccuracy) {
+    public void setHorizontalAccuracy(Double horizontalAccuracy) {
         this.horizontalAccuracy = horizontalAccuracy;
     }
 
@@ -121,19 +121,19 @@ public class TrajectoryEntity extends BaseEntity {
         this.latitude = latitude;
     }
 
-    public BigDecimal getSpeed() {
+    public Double getSpeed() {
         return speed;
     }
 
-    public void setSpeed(BigDecimal speed) {
+    public void setSpeed(Double speed) {
         this.speed = speed;
     }
 
-    public BigDecimal getVerticalAccuracy() {
+    public Double getVerticalAccuracy() {
         return verticalAccuracy;
     }
 
-    public void setVerticalAccuracy(BigDecimal verticalAccuracy) {
+    public void setVerticalAccuracy(Double verticalAccuracy) {
         this.verticalAccuracy = verticalAccuracy;
     }
